backends/nsfw_backend/requirements.txt
transformers
torch
pillow
numpy
timm
